fix: eslint and ts config
