+++ src/my-app/package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "start": "vite --port 3000",
     "build": "vite build && tsc",
-    "prebuild": "dfx generate backend"
+    "prebuild": "dfx generate backend",
+    "lint": "eslint ."
   },
   "dependencies": {
     "@dfinity/agent": "^2.4.0",
@@ -51,6 +52,7 @@
     "zod": "^3.24.1"
   },
   "devDependencies": {
+    "@eslint/js": "^9.23.0",
     "@testing-library/dom": "^10.4.0",
     "@testing-library/react": "^16.2.0",
     "@types/node": "^20",
@@ -59,9 +61,14 @@
     "@vitejs/plugin-react": "^4.3.4",
     "autoprefixer": "^10.4.21",
     "dotenv": "^16.4.7",
+    "eslint": "^9",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-react-refresh": "^0.4.19",
+    "globals": "^16.0.0",
     "jsdom": "^26.0.0",
     "postcss": "^8",
     "typescript": "^5.8.2",
+    "typescript-eslint": "^8.28.0",
     "vite": "^6.1.0",
     "vitest": "^3.0.5",
     "web-vitals": "^4.2.4"
