+++ src/my-app/src/components/edit-user-dialog.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { Button } from '@/components/ui/button';
 import { useState } from 'react';
 import {
