+++ src/my-app/src/components/ui/badge.tsx
@@ -35,4 +35,5 @@ function Badge({ className, variant, ...props }: BadgeProps) {
   );
 }
 
+// eslint-disable-next-line react-refresh/only-export-components
 export { Badge, badgeVariants };
