+++ src/my-app/src/components/wheel/container.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { cn } from '@/lib/utils';
 
 type WheelContainerProps = React.HTMLAttributes<HTMLDivElement>;
