+++ src/my-app/src/hooks/use-activity.ts
@@ -1,5 +1,3 @@
-'use client';
-
 import { useQuery, type UseQueryResult } from '@tanstack/react-query';
 import type {
   Err,
