+++ src/my-app/src/hooks/use-create-wheel-asset.ts
@@ -1,5 +1,3 @@
-'use client';
-
 import { useAuth } from '@/contexts/auth-context';
 import { extractOk } from '@/lib/api';
 import { candidOpt, toastError } from '@/lib/utils';
