+++ src/my-app/src/hooks/use-create-wheel-prize-extraction.ts
@@ -1,5 +1,3 @@
-'use client';
-
 import { useMutation } from '@tanstack/react-query';
 import { Principal } from '@dfinity/principal';
 import { useAuth } from '@/contexts/auth-context';
