+++ src/my-app/src/hooks/use-poll-extraction.ts
@@ -1,5 +1,3 @@
-'use client';
-
 import { useAuth } from '@/contexts/auth-context';
 import { useWheelPrizes } from '@/hooks/use-wheel-prizes';
 import { useQuery } from '@tanstack/react-query';
