+++ src/my-app/src/hooks/use-toast.ts
@@ -1,5 +1,3 @@
-'use client';
-
 // Inspired by react-hot-toast library
 import * as React from 'react';
