+++ src/my-app/src/hooks/use-update-wheel-asset.ts
@@ -1,5 +1,3 @@
-'use client';
-
 import { useAuth } from '@/contexts/auth-context';
 import { extractOk } from '@/lib/api';
 import { candidOpt, toastError, toCandidEnum } from '@/lib/utils';
