+++ src/my-app/src/hooks/use-wheel-asset-tokens.ts
@@ -1,5 +1,3 @@
-'use client';
-
 import { useAuth } from '@/contexts/auth-context';
 import {
   isWheelAssetToken,
