+++ src/my-app/src/main.tsx
@@ -1,23 +1,19 @@
 import { StrictMode } from 'react';
 import ReactDOM from 'react-dom/client';
-import { RouterProvider, createRouter } from '@tanstack/react-router';
 import {
   QueryCache,
   QueryClient,
   QueryClientProvider,
 } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-
-// Import the generated route tree
-import { routeTree } from './routeTree.gen';
-
-import './styles.css';
 import reportWebVitals from './reportWebVitals.ts';
-import { AuthProvider, useAuth } from '@/contexts/auth-context';
+import { AuthProvider } from '@/contexts/auth-context';
 import { toast } from '@/hooks/use-toast';
 import { renderError } from '@/lib/utils';
 import { Toaster } from '@/components/ui/toaster';
 import { printVersionToConsole } from '@/lib/utils';
+import App from './App.tsx';
+import './styles.css';
 
 printVersionToConsole();
 
@@ -34,30 +30,6 @@ const queryClient = new QueryClient({
     },
   }),
 });
-
-// Create a new router instance
-const router = createRouter({
-  routeTree,
-  context: {
-    auth: undefined!,
-  },
-  defaultPreload: 'intent',
-  scrollRestoration: true,
-  defaultStructuralSharing: true,
-  defaultPreloadStaleTime: 0,
-});
-
-// Register the router instance for type safety
-declare module '@tanstack/react-router' {
-  interface Register {
-    router: typeof router;
-  }
-}
-
-function App() {
-  const authContext = useAuth();
-  return <RouterProvider router={router} context={{ auth: authContext }} />;
-}
 
 // Render the app
 const rootElement = document.getElementById('app');
